WIP DevSim:Initial shell code to handle GPDP2 ext
This adds some placeholder code to show changes that need to be made to
add support for extended property queries made via
vkGetPhysicalDeviceProperties2KHR().

There are some updates required to the json file reading that weren't
immediately obvious to me so made this change as a start and hoping
someone else can finish it.