Android: do fuzzy image comparisons
On Android devices, we get frequent failures from image comparisons where the
images are not identical.  This code changes the comparison so it uses
an ImageMagick RMSE (Root Mean Squared Error) comparison against a threshold,
instead of an absolute image match.  The PASS and FAIL messages include
the actual RMSE metric value.

Also, save separate images for each device; right now only the images from the
last device run are saved, which makes it difficult to triage a failure.