json_gen: Append trailing slashes to filenames
The json_gen layer queries the directory to store json/spv dumps
from the envvar VK_JSON_FILE_PATH.

While the json dumps were being stored in the correct dir pointed
to by VK_JSON_FILE_PATH, the spv and object reservation dumps were
being stored at one level above the directory specified in
VK_JSON_FILE_PATH if the value in VK_JSON_FILE_PATH did not
contain a trailing slash

This change unconditionally adds a trailing slash to the paths
used for storing spv and objection reservation dumps.

While at it, this change also adds a function getBaseDirectoryPath()
that is invoked once in the ctor of PipelineData to get the base path
along with a call to getProcessName() to get the name of the process to
avoid invoking it multiple times.